aws_code_deploy_configure_files

+++ scripts/stop_and_remove_images.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 pre_containers = $(docker ps -a -q);
+if [$pre_containers == "" || $pre_containers == null] exit 0
 docker stop "$pre_containers";
 docker rm "$pre_containers"
